Moved sc2 maps processing functions
